test: move router_strategy tests to CI-run directory (tests/test_litellm/router_strategy/)
