+++ tests/local_testing/test_tpm_rpm_routing_v2.py
@@ -27,9 +27,6 @@ from litellm.caching.caching import DualCache
 from litellm.router_strategy.lowest_tpm_rpm_v2 import (
     LowestTPMLoggingHandler_v2 as LowestTPMLoggingHandler,
 )
-from litellm.router_strategy.lowest_tpm_rpm import (
-    LowestTPMLoggingHandler as LowestTPMLoggingHandler_v1,
-)
 from litellm.utils import get_utc_datetime
 from create_mock_standard_logging_payload import create_standard_logging_payload
 
@@ -38,20 +35,6 @@ from create_mock_standard_logging_payload import create_standard_logging_payload
 """
 - Given 2 deployments, make sure it's shuffling deployments correctly.
 """
-
-
-def test_lowest_tpm_rpm_v1_model_info_none():
-    """log_success_event should not crash when model_info is None (v1 handler)."""
-    logger = LowestTPMLoggingHandler_v1(router_cache=DualCache())
-    kwargs = {
-        "litellm_params": {
-            "metadata": {"model_group": "gpt-4"},
-            "model_info": None,
-        }
-    }
-    logger.log_success_event(
-        kwargs=kwargs, response_obj={}, start_time=time.time(), end_time=time.time()
-    )
 
 
 def test_tpm_rpm_updated():
